+++ Changelog.md
@@ -1,3 +1,8 @@
+### 5.9    
+Remove .md as a scenario property file    
+Bug fix  
+Add Error support in case of duplicated scenario properties are recursively loaded from directory    
+
 ### 5.8    
 Bug fix  
 Refactoring